Implement and refactor Actors, better logging
This commit refactors the actors into their own file. Additionally, the
`gpio_relay` actor is added.

A new function is introduced to initialize all actors at once which
simplifies the code and allows for creating all actors from the
configuration file.

As a last part, the logging was improved to allow better logging for
systemd. This keeps the redundant information away.